Fix issue 93: TSVN fails to handle svn:externals correctly upon commit.
When collecting all the statuses, also check the uuid of each file/folder. Non-equal uuids mean we have an svn:external from another repository and can't commit those in one step. So exclude those with a different uuid than the first we got (usually that from the parent dir), but show the user a warning dialog telling that there are modifications in one or more svn:external folders which need a separate commit.

git-svn-id: http://tortoisesvn.googlecode.com/svn/trunk@853 4cfb86a4-2c7a-11df-bd14-c1e331e6ceac