Improved shell-update mechanism.  Previously, shell-update requests were being done once-per-target.  This caused a bad interaction with the shell-extension status cache, where first update would load the cache, and subsequent updates would get stale, cached data.
I have added a named event to signal from TortoiseProc to the shellext, to tell it to invalidate its cache.  I have added a collection-class to TortoiseProc which queues-up shell-update requests until either a timeout expires, or the app exits, at which point all the updates are done in one go.

git-svn-id: http://tortoisesvn.googlecode.com/svn/trunk@2278 4cfb86a4-2c7a-11df-bd14-c1e331e6ceac