Start the process of making the command line more consistent in its use of the /path: argument.
This rev removes all intermediate use of path-list tempfiles throughout the app, so that code which once took a path to a tempfile now takes a CTSVNPathList.  Added a commandInfo table to TortoiseProc.cpp to allow common code to do all the dealing with the /path: argument before going into the individual per-command stuff.

Rather a bigger mod that I would have liked, but hopefully it removes all the cases where one could never tell if sPath was a real path, or just the name of a tempfile.

This revision shouldn't make any difference to callers of TortoiseProc.exe


git-svn-id: http://tortoisesvn.googlecode.com/svn/trunk@2409 4cfb86a4-2c7a-11df-bd14-c1e331e6ceac