Show only those properties in the combobox which actually can be used on the selected item.
I.e. if a file is selected in the Subversion properties tab, then only add those properties which can be used on files, and if a folder is selected only show those which can be used on folders.
If more than one item (file or folder) is selected, all properties are added to the combobox.

git-svn-id: http://tortoisesvn.googlecode.com/svn/trunk@2627 4cfb86a4-2c7a-11df-bd14-c1e331e6ceac