Restructure the DUG. Issue #16
This is an attempt to make the DUG clearer and easier to use.

- Rework 'Handling TSVN' and 'Authentication' into a new 'Getting Started'.
- Rename 'Get/Set Properties of Files and Folders' to 'Project Properties'.
- Add a picture of the right drag menu.
- Reorder the sections closer to the order they are likely to be encountered.

Note: The German files have been modified to show the new structure,
but some translations are needed in:
- de/dug_general.xml
- de/dug_propertypage.xml

RightDragMenu.png needs to be recaptured in XP, and in German.

The German docs do not yet have a RevGraph.xml file.

git-svn-id: http://tortoisesvn.googlecode.com/svn/trunk@2802 4cfb86a4-2c7a-11df-bd14-c1e331e6ceac