Include the thesaurus and spell checker dictionary in the installer and the language packs if they're found in the ..\Common\Spell_Thes folder when building.
Also, don't include the crashrpt.dll in the installer if the file MYBUILD isn't in the TSVN root folder - this prevents crashreports sent to me for which I don't have the source and/or the debug symbols.

The next version of Wix (2.0.2912) generates warnings if the 'NAME' tag has a filename which ends in '~digit'. So change all those names to something else - it doesn't really matter what since we don't install TSVN on DOS.
Note: we can't use 2.0.2912 yet, because of bug #1207235 in WiX: 
http://sourceforge.net/tracker/index.php?func=detail&aid=1207235&group_id=105970&atid=642714


git-svn-id: http://tortoisesvn.googlecode.com/svn/trunk@3541 4cfb86a4-2c7a-11df-bd14-c1e331e6ceac