Removed radiobuttons for choosing everyone/currentuser profile, and applied "sensible default".
If the user is an admin (and VersionNT>=400 - ie, NT4 and above), install to the all users profile.  Otherwise, install to the current user's profile.  This can (I assume) be overridden by specifying a value for ALLUSERS on the command-line - ie, msiexec /i Setup.msi ALLUSERS=<something other than 2> kind of thing.

git-svn-id: http://tortoisesvn.googlecode.com/svn/trunk@4144 4cfb86a4-2c7a-11df-bd14-c1e331e6ceac