Implement issue #210:
Use different icons for the TortoiseSVN context menu in the explorer, depending on what was selected to show the context menu.

* if a folder is selected, show a Tortoise with a folder icon
* if a file is selected, show a Tortoise with a file icon
* if a shortcut is selected, show a Tortoise with a link icon
* if multiple items are selected, show a Tortoise with a file-plus icon

To avoid having two (or more) TSVN entries for folders in the context menu, we check if we already inserted our entry in the menu and don't insert another one in that case.

git-svn-id: http://tortoisesvn.googlecode.com/svn/trunk@5621 4cfb86a4-2c7a-11df-bd14-c1e331e6ceac