Patch from Dirk (vss2svn at nogga.de):
there is a small bug in the ResText utility that prevents DLGTEMPLATE to be converted correctly.

The definition of the DLGTEMPLATE structure (taken from the MSDN) is:

typedef struct {
   DWORD style;
   DWORD dwExtendedStyle;
   WORD cdit;
   short x;
   short y;
   short cx;
   short cy;
} DLGTEMPLATE, *LPDLGTEMPLATE;

and the style parameter was already parsed in the CountMemReplaceDialogResource. The original code tried to reparse the style parameter after the dwExtendedStyle parameter.


git-svn-id: http://tortoisesvn.googlecode.com/svn/trunk@5625 4cfb86a4-2c7a-11df-bd14-c1e331e6ceac